ibks_twitter/requirements.txt
django-simple-captcha-0.6.0
django-email-verification-0.3.3